Handler Members CSS fixes
Done:
Fixed handler.js, members.njk, icons.css to display the members and heads.
Tweaked the handler to only display the current members since we are going to use the legacy format.
To-do:
Update pfp naming format using id as a reference instead of name since we have 2 Ayush Kumar.
Add the pfp for new guys.